hnsw: turn neurips23 test into benchmark
