+++ adapters/repos/db/vector/hnsw/.gitignore
@@ -1,3 +1,5 @@
 recall_vectors.json
 recall_queries.json
 recall_truths.json
+datasets/ann-benchmarks/
+datasets/big-ann-benchmarks/
